Change package name to rodi_robot
The package name has been changed to rodi_robot to be more aligned with
the ROS package naming advice (http://www.ros.org/reps/rep-0144.html)
and to avoid using the "ws" suffix, since in ROS this usually refers to
a workspace so is confusing to use it in a package name.

Also change the rodi_ws_node to rodi_robot_node for consistency.

"There are only two hard things in Computer Science: cache invalidation
and naming things." -- Phil Karlton