transport: Reduce the httplib timeout to 0.5 seconds
Under a heavy load, RoDI isn't able to process the HTTP requests and the
connections timeouts. But waiting for a second is too much and makes the
node to be unresponsive for a second, until the HTTP requests timeouts.

If RoDI will not be able to process the request, it's better to make the
request to timeout early to allow the next command to be processed.